pyproject.toml
[project]
name = "semvergit"
description = "Manage your project's version numbers."
requires-python = ">=3.8"
authors = [
    {name = "Roy Moore", email="roy@moore.co.il"}
]
dynamic = ["version"]
dependencies = [
    "gitpython==3.1.40",
    "loguru==0.7.2",
    "click==8.1.7",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "pylint",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "mypy",
]

[project.urls]
homepage = "https://github.com/Tranquility2/semvergit"
repository = "https://github.com/Tranquility2/semvergit"

[tool.isort]
line_length = 120

[tool.black]
line-length = 120

[tool.pylint.main]
fail-under = 10.0

[tool.pylint.format]
max-line-length = 120

[tool.coverage.run]
omit = [
    "semvergit/tests/*",
    "src/semvergit/__main__.py",
]

[tool.coverage.report]
fail_under = 100

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_defs = true
implicit_reexport = false
warn_unused_ignores = true
warn_redundant_casts = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
show_traceback = true
